requirements.txt
fastapi~=0.87.0
gunicorn
uvicorn
pytest
pre-commit
requests
celery~=5.2.7
aiofiles
redis
cassandra-driver~=3.25.0
dateparser
pandas
requests_html
selenium
python-dotenv~=0.21.0
python-slugify
fake-useragent
pydantic~=1.10.2
loguru~=0.6.0